Make sure the compiler won't pre-calculate with chars to 0 - cast to float (cppcheck)

+++ src/cmap.cxx
@@ -120,11 +120,11 @@ typedef unsigned char uchar;
 void background(uchar r, uchar g, uchar b) {
   // replace the gray ramp so that color 47 (by default 2/3) is this color
   if (!r) r = 1; else if (r==255) r = 254;
-  double powr = log(r/255.0)/log((FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
+  double powr = log(r/255.0)/log(((float) FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
   if (!g) g = 1; else if (g==255) g = 254;
-  double powg = log(g/255.0)/log((FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
+  double powg = log(g/255.0)/log(((float) FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
   if (!b) b = 1; else if (b==255) b = 254;
-  double powb = log(b/255.0)/log((FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
+  double powb = log(b/255.0)/log(((float) FL_GRAY-FL_GRAY_RAMP)/(FL_NUM_GRAY-1.0));
   for (int i = 0; i < FL_NUM_GRAY; i++) {
     double gray = i/(FL_NUM_GRAY-1.0);
     cmap[i+FL_GRAY_RAMP][0] = uchar(pow(gray,powr)*255+.5);
